Revert non-structural changes from namespace rename
