+++ arc_agi_3_agents/templates/README.md
