+++ arc_agi_3_agents/templates/langgraph_thinking/__init__.py
