+++ arc_agi_3_agents/templates/langgraph_thinking/vision.py
