+++ arc_agi_3_agents/templates/random_agent.py
